@@ -1,17 +1,60 @@
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.IdentityModel.Tokens;
+using Microsoft.OpenApi.Models;
 using SalePoint.Primitives.Interfaces;
 using SalePoint.Repository;
 using System.Text;
 
+var MyAllowSpecificOrigins = "_myAllowSpecificOrigins";
+
 var builder = WebApplication.CreateBuilder(args);
+
+builder.Services.AddCors(options =>
+{
+    options.AddPolicy(name: MyAllowSpecificOrigins,
+        policy =>
+        {
+            policy.WithOrigins("http://localhost:3000");
+            //policy.AllowAnyOrigin();
+            policy.WithMethods("GET", "POST", "PUT", "DELETE");
+            policy.WithHeaders("*");
+        });
+});
 
 // Add services to the container.
 
 builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
+
+builder.Services.AddSwaggerGen(opt =>
+{
+    opt.SwaggerDoc("v1", new OpenApiInfo { Title = "SalePointApi", Version = "v1" });
+    opt.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
+    {
+        In = ParameterLocation.Header,
+        Description = "Please enter token",
+        Name = "Authorization",
+        Type = SecuritySchemeType.Http,
+        BearerFormat = "JWT",
+        Scheme = "bearer"
+    });
+    opt.AddSecurityRequirement(new OpenApiSecurityRequirement
+    {
+        {
+            new OpenApiSecurityScheme
+            {
+                Reference = new OpenApiReference
+                {
+                    Type=ReferenceType.SecurityScheme,
+                    Id="Bearer"
+                }
+            },
+            Array.Empty<string>()
+        }
+    });
+});
+
 builder.Services.AddSingleton<IProductRepository, ProductRepository>();
 builder.Services.AddSingleton<IDepartmentRepository, DepartmentRepository>();
 builder.Services.AddSingleton<IMeasurementUnitRepository, MeasurementUnitRepository>();
@@ -19,6 +62,7 @@ builder.Services.AddSingleton<IUserRepository, UserRepository>();
 builder.Services.AddSingleton<IRolRepository, RolRepository>();
 builder.Services.AddSingleton<ICashRegisterRepository, CashRegisterRepository>();
 builder.Services.AddSingleton<ISaleRepository, SaleRepository>();
+builder.Services.AddSingleton<ITicketRepository, TicketRepository>();
 
 builder.Services.AddAuthentication(
     option =>
@@ -28,7 +72,7 @@ builder.Services.AddAuthentication(
     }).
    AddJwtBearer(o =>
    {
-       var key = Encoding.UTF8.GetBytes(builder.Configuration["JWT:Key"]);
+       var key = Encoding.UTF8.GetBytes(builder.Configuration["JWT:Key"]!);
        o.SaveToken = true;
        o.TokenValidationParameters = new TokenValidationParameters
        {
@@ -45,13 +89,16 @@ builder.Services.AddAuthentication(
 var app = builder.Build();
 
 // Configure the HTTP request pipeline.
-if (app.Environment.IsProduction())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
+//if (app.Environment.IsProduction())
+//{
+//    app.UseSwagger();
+//    app.UseSwaggerUI();
+//}
+app.UseSwagger();
+app.UseSwaggerUI();
 app.UseHttpsRedirection();
+app.UseCors(MyAllowSpecificOrigins);
+//app.UseCors(x => x.WithOrigins("*").WithMethods("GET", "POST", "PUT", "DELETE").WithHeaders("*"));
 app.UseAuthentication();
 app.UseAuthorization();
 
