@@ -2,8 +2,8 @@
 {
     public record Access
     {
-        public string UserName { get; set; }
+        public required string UserName { get; set; }
 
-        public string Pass { get; set; }
+        public required string Pass { get; set; }
     }
 }
